Basic functionality Widget tests
Rest App functionality is pretty basic. It has a test field to track the timer seconds.
To buttons for adding or substracting time from the timer.
And a button to start/stop the timer.

The default timer is 20 minutes. And the amount for add/substract is 5 minutes.

When the timer has started, add and substract buttons are hidden.