App service started on demand
When app intitializes it start and connects (bound) to the service.
It shows a SnackBar whether it could connect or not to the service.
If it wasn't able to connect to the service, the start/stop button is
disabled.

After connected to the service it gets the server current seconds.
If it 0 it means the timer wasn't running, therefore timer shows
stopped.